Extract `FeedImagesMapper` into its own file
